move reset constants to consts
